Added InterService communication with FeignClient(No Eureka)

+++ CartService/src/main/java/com/example/cart/cartService/model/CartItem.java
@@ -20,7 +20,7 @@ public class CartItem {
     @Column(name = "item_name")
     private String itemName;
     @Column(name = "item_qty")
-    private Long itemQuantity;
+    private Double itemQuantity;
     @Column(name = "item_price")
     private Double itemPrice;
     @Column(name = "restaurant_id")
